[bugfix] Corrected DOMImplementation#hasFeature("XML", "")
